requirements.txt
boto3>=1.37.12
cryptography==43.0.1
dj-database-url==2.3.0
Django<=5.1.6
django-environ==0.12.0
django-heroku==0.3.1
django-allauth==65.4.1
django-storages==1.14.5
requests==2.31.0
PyJWT==2.9.0
gunicorn>=20.1.0
Pillow==11.1.0